Refactor to remove use of namespace in types. Add cpu_usage_us, net_usage_words, parent, signatures to Action. Add missing queryParams to get_links.
