+++ src/endpoints.ts
@@ -1,6 +1,6 @@
 import {APIClient} from '@wharfkit/antelope'
-import {HyperionV1APIClient} from './endpoints/v1'
-import {HyperionV2APIClient} from './endpoints/v2'
+import {HyperionV1APIClient} from './endpoints/v1/v1'
+import {HyperionV2APIClient} from './endpoints/v2/v2'
 
 export class HyperionAPIClient {
     public v1: HyperionV1APIClient
