+++ test/data/9a482202a8200a1195142218f306079b32dc4764.json
@@ -1,0 +1,19 @@
+{
+    "request": {
+        "path": "https://wax.blokcrafters.io/v2/history/get_creator?account=eosio",
+        "params": {
+            "method": "GET"
+        }
+    },
+    "status": 200,
+    "json": {
+        "query_time_ms": 0.262,
+        "cached": true,
+        "account": "eosio",
+        "creator": "__self__",
+        "timestamp": "2019-06-05T12:00:00.000",
+        "block_num": 1,
+        "trx_id": ""
+    },
+    "text": "{\"query_time_ms\":0.262,\"cached\":true,\"account\":\"eosio\",\"creator\":\"__self__\",\"timestamp\":\"2019-06-05T12:00:00.000\",\"block_num\":1,\"trx_id\":\"\"}"
+}
