+++ test/tsconfig.json
@@ -6,7 +6,10 @@
         "module": "commonjs",
         "target": "es6",
         "types": ["mocha", "node"],
-        "baseUrl": ".."
+        "baseUrl": "..",
+        "paths": {
+            "$lib": ["."]
+        },
     },
     "include": ["*.ts", "**/*.ts"]
 }
